refactor: change structure of the project based on FSD principles
