+++ src/pages/main.tsx
@@ -1,7 +1,6 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import './index.css';
-import App from './App.tsx';
+import App from '../app/App.tsx';
 
 createRoot(
   document.getElementById('root') ?? document.createElement('div')
